+++ finder/src/main/java/com/finder/finder/service/FinderService.java
@@ -2,5 +2,5 @@ package com.finder.finder.service;
 
 public interface FinderService {
 
-    Integer findNumber(String filePath, int n);
+    Long findNumber(String filePath, int n);
 }
